# packages/bzx-api/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a public API for bZx protocol, Fulcrum and Torque.",
    "version": "1.0.0",
    "title": "Fulcrum API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.bzx.network",
  "basePath": "/v1",
...
  "schemes": ["https"],
  "paths": {
...
    "/borrow-deposit-estimate": {
      "get": {
        "tags": ["Torque"],
        "summary": "Collateral Estimate",
        "description": "Estimates how much of collateral you have to deposit to borrow AMOUNT of asset",
        "parameters": [
          {
            "name": "borrow_asset",
            "in": "query",
            "required": "true",
            "type": "string",
            "description": "Borrow asset"
          },
          {
            "name": "collateral_asset",
            "in": "query",
            "required": "true",
            "type": "string",
            "description": "Collateral asset"
          },
          {
            "name": "amount",
            "in": "query",
            "required": "true",
            "type": "integer",
            "description": "Loan amount"
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/tvl-history": {
      "get": {
        "tags": ["History Data"],
        "summary": "Get tvl history data",
        "description": "",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": "true",
            "type": "integer",
            "description": "Start Date (timestamp)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": "true",
            "type": "integer",
            "description": "End Date (timestamp)"
          },
          {
            "name": "points_number",
            "in": "query",
            "required": "true",
            "type": "integer",
            "description": "number of desired datapoints output"
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/asset-stats-history": {
      "get": {
        "tags": ["History Data"],
        "summary": "Get history data for the asset (tvl, apr, utilization)",
        "description": "",
        "parameters": [
          {
            "name": "asset",
            "in": "query",
            "required": "true",
            "type": "string",
            "description": "Asset"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": "true",
            "type": "integer",
            "description": "Start Date (timestamp)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": "true",
            "type": "integer",
            "description": "End Date (timestamp)"
          },
          {
            "name": "points_number",
            "in": "query",
            "required": "true",
            "type": "integer",
            "description": "number of desired datapoints output"
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/asset-history-price": {
      "get": {
        "tags": ["History Data"],
        "summary": "Get asset's swap to USD price at some timestamp",
        "description": "",
        "parameters": [
          {
            "name": "asset",
            "in": "query",
            "required": "true",
            "type": "string",
            "description": "Asset"
          },
          {
            "name": "date",
            "in": "query",
            "required": "true",
            "type": "integer",
            "description": "Date (timestamp)"
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}